+++ mountpoint-s3-client/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Other changes
 
+* Add support for CRC64-NVME checksum algorithm. ([#1235](https://github.com/awslabs/mountpoint-s3/pull/1235))
 * Add support for overriding the number of threads used by the Event Loop.
   ([#1240](https://github.com/awslabs/mountpoint-s3/pull/1240/))
 * The ECS credentials provider now performs retries in the event of some failures. ([awslabs/aws-c-auth#259](https://github.com/awslabs/aws-c-auth/pull/259/))